chore: Rename log to avoid clash with logging
